requirements.txt
streamlit
openai>=1.30.0
pandas
python-dotenv